api v1 orders create

+++ database/migrations/tenant/2023_05_17_031835_create_products_table.php
@@ -15,9 +15,9 @@ return new class extends Migration {
             $table->integer('availability');
             $table->string('title');
             $table->text('description')->nullable();
-            $table->float('trade_price')->nullable();
-            $table->float('price');
-            $table->float('discount_price')->nullable();
+            $table->decimal('trade_price')->nullable();
+            $table->decimal('price');
+            $table->decimal('discount_price')->nullable();
             $table->foreignId('preview_id')->nullable();
             $table->string('sku')->nullable();
             $table->timestamps();
