+++ database/migrations/tenant/2023_05_19_203444_create_order_costs_table.php
@@ -18,7 +18,7 @@ return new class extends Migration {
                 ->on('orders')
                 ->onDelete('cascade');
             $table->string('title');
-            $table->float('sum');
+            $table->decimal('sum');
             $table->text('comment')->nullable();
             $table->timestamps();
         });
